Moved ffplay into video module
