+++ subtitler/cli/main_gcc.cpp
@@ -6,7 +6,7 @@
 #include <glog/logging.h>
 
 #include "subtitler/subprocess/subprocess_executor.h"
-#include "subtitler/play_video/ffplay.h"
+#include "subtitler/video/player/ffplay.h"
 #include "subtitler/cli/commands.h"
 #include "subtitler/cli/io/input.h"
 
@@ -119,10 +119,10 @@ int main(int argc, char **argv) {
         ofs << "";
     }
 
-    auto input_getter = std::make_unique<cli::io::NarrowInputGetter>(std::cin);
+    cli::Commands::Paths paths{FLAGS_video_path, FLAGS_output_subtitle_path};
     auto executor = std::make_unique<subprocess::SubprocessExecutor>();
     auto ffplay = std::make_unique<play_video::FFPlay>(FLAGS_ffplay_path, std::move(executor));
-    cli::Commands::Paths paths{FLAGS_video_path, FLAGS_output_subtitle_path};
+    auto input_getter = std::make_unique<cli::io::NarrowInputGetter>(std::cin);
     cli::Commands commands{paths, std::move(ffplay), std::move(input_getter), std::cout};
 
     try {
